actualizacion

+++ Capitulo08/Matematicas/ConversorDeSistemas.java
@@ -30,9 +30,9 @@
   */
   public static long conversorDecimalBinarioLento(long x){
     int counter = 0;
-    int bits = 0;
-    int restaCiclica = 0;
-    int result = 0;
+    long bits = 0;
+    long restaCiclica = 0;
+    long result = 0;
     while(bits < x){
       bits = Matematicas.OperacionesEnteros.potencia(2, counter);
       if(bits < x){
